fix(QuickAddWord): Add enter function
