+++ src/App.tsx
@@ -4,9 +4,9 @@ import WordList from "./components/WordList";
 import data from "./data.json";
 import BasicInput from "./components/BasicInput";
 import { flatten } from "lodash";
+import QuickAddWordForm from "./components/QuickAddWordForm";
 
 function App() {
-  const quickAddWordRef = React.useRef(null);
   const [_data, setFilteredList] = React.useState(data);
   const [basicInputValues, setBasicInputValues] = React.useState(new Array(5).fill(""));
   const [locksValue, setLocksValue] = React.useState(new Array(5).fill(false));
@@ -143,11 +143,6 @@ function App() {
     );
   }
 
-  function handleQuickAddWord() {
-    handleWordClick(quickAddWordRef.current.value?.toLowerCase());
-    quickAddWordRef.current.value = "";
-  }
-
   function handleChangeExtraCharactersToExclude(newValue) {
     setExtraCharactersToExcludeRaw(newValue);
     handleWordlistChange(
@@ -178,28 +173,7 @@ function App() {
             })}
           </code>
         </div>
-        <div className="pl-4 pr-4 pb-4 flex h-12 items-center justify-center">
-          <input
-            ref={quickAddWordRef}
-            data-testid="quickAddWordInput"
-            type="text"
-            placeholder="Quick add word"
-            max={5}
-            className="block
-            w-full
-            rounded-md
-            bg-gray-100 dark:bg-gray-600 dark:text-white
-            border-transparent
-            focus:border-gray-500 focus:bg-white dark:focus:bg-gray-700 focus:ring-0"
-          />
-          <button
-            data-testid="quickAddWordButton"
-            onClick={handleQuickAddWord}
-            className="bg-red-700 text-white p-3 w-auto rounded-lg hover:bg-red-900 focus:outline-none focus:ring shadow-md hover:shadow-none transition-all duration-300"
-          >
-            +
-          </button>
-        </div>
+        <QuickAddWordForm onSubmit={handleWordClick} />
         <div className="pl-4 pr-4">
           <input
             data-testid="extraCharactersToExclude"
